Introduced important add-on: exact numbers / computations, as opposed to current implementation based on class "FixedPointNumber" in SchnorxoLib, which allows for high/arbitrary-precision computations, but not for exact ones).
(read-only, for the moment).

(Based on SchnorxoLib V. 0.2 and
class "BigFraction" in Apache Commons Numbers. ==> new dependency)

 - Package "read":
   o GnuCashTrxSplit(Impl): New methods getValueRat() and getQuantityRat()
     ("rat" for "rational number") that return the according values as
     fractions / rational numbers.
   o GnuCashAccount(Impl) GnuCashTransaction(Impl): new methods getBalanceXYZRat()
     (symmetric to already existing ones) that return balance of account /
     transaction as *exactly computed* number (i.e., as fraction).
   o New test cases for that.
   o These new BigFraction-twins of various methods introduce some code redundancies.
     In order to better manage them / make code more readable and maintainable, I
     have transferred all getBalanceXYZ() methods in GnuCashAccountImpl to two
     new helper classes:
     AccountBalanceHelper_FN (old code) and
     AccountBalanceHelper_BF (new code).
     *Not* so for GnuCashTransaction(Split)Impl.
 - Package "currency": Introduced new methods in ComplexPriceTable, SimplePriceTable
   and its derivatives. (::TODO: Some methods cannot be translated 1-by-1 into their
   BigFraction twins, will have to to a minor redesign here. Cf. comments in code)
 - module-info: introduced new dependency.

Rationale:

One might ask: "Isn't that a bit exaggerated? Granted, there are some things /
computations that get a bit awkward with FixedPointNumbers (i.e. BigDecimal,
interally), like the famous examples 1/3 or 2/3 etc. -- but still: This is not
rocket science, we still have just an accounting application here! 8, 10 or 12
digits, whatever, should be more than enough for it, dont' you think so? You are
introducing more complexity for merely theoretical cases!"

Well -- yes and no. You see: The GnuCash developers (and the ones of KMyMoney,
the sister project as well) must have had a thought or two over this: After
all, they chose to represent *every* number in their resp. programs as a fraction.
Guess why! Now, it simply seems weird to me to have a Java lib / Java programs to
read and manipulate this data -- and do computations with it -- that do *not*
adequately take this design choice into account, even if -- granted -- that means
an unnecessary overhead in 99.99% of all cases that are practically relevant (and
that is one of the reasons why I don't like the SchnorxoLib's class "FixedPointNumber",
written by the original author, and want to get rid of it / reduce the dependency
to it in the long run). (And by the way: what about the other 0.01%? Should I
just ignore them?).

Sorry, Marcus, I value the work you have done and your donation to the community,
but FixedPointNumber simply was a bad decision. You should have gone with the
foundational design decision of GnuCash's developers (i.e. using fractions and
computing exactly with them internally) instead of working against it. This is a
step towards going with them.